requirements.txt
feedparser
requests
beautifulsoup4
python-telegram-bot==20.3
deep-translator
